Work around some cells which do not load
In Unified layout, especially with feeds containing images, some cells
do not load, unless the user scrolls forth and back or resize the window
to force these cells to reload.
The main cause is that, for some unknown reason, some html files are
missing in action while they seem to have been correctly created.
We detect the situation through webView:didFailProvisionalNavigation:withError:
and provoke a reload of the cell. 

Might solve issue #1205